InputBrowser did not compile on OS/X (possibly gcc bug). Win32 after a ready() would block in wait() waiting for a second event Changed fltk::version() to return double instead of float
git-svn-id: file:///fltk/svn/trunk@4629 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/run.h
@@ -35,7 +35,7 @@ namespace fltk {
 
 /*! \addtogroup startup
   \{ */
-FL_API float version();
+FL_API double version();
 FL_API void display(const char*);
 FL_API int arg(int, char**, int&);
 FL_API int args(int, char**, int&, int (*)(int,char**,int&) = 0);
